+++ roothash/api/src/lib.rs
@@ -1,0 +1,12 @@
+extern crate ekiden_common_api;
+extern crate ekiden_scheduler_api;
+extern crate futures;
+extern crate grpcio;
+extern crate protobuf;
+
+mod generated;
+
+use ekiden_scheduler_api as scheduler;
+
+pub use generated::roothash::*;
+pub use generated::roothash_grpc::*;